Move security-related classes to their own package
