+++ application/src/main/java/com/musicforall/config/security/CustomAuthenticationFailureHandler.java
@@ -1,4 +1,4 @@
-package com.musicforall.config;
+package com.musicforall.config.security;
 
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.authentication.SimpleUrlAuthenticationFailureHandler;
