+++ application/src/main/java/com/musicforall/config/security/CustomAuthenticationProvider.java
@@ -1,4 +1,4 @@
-package com.musicforall.config;
+package com.musicforall.config.security;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.AuthenticationProvider;
